feat(settings): split Preferences and group nav into Personal/Workspace
