+++ packages/frontend/src/composable/data-queries/user-settings.ts
@@ -28,6 +28,7 @@ export const useUserSettings = (
     invalidate,
     mutate: mutation.mutate,
     mutateAsync: mutation.mutateAsync,
+    isUpdating: mutation.isPending,
     ...query,
   };
 };
