+++ packages/frontend/src/i18n/index.ts
@@ -57,7 +57,9 @@ const chunkRegistry: ChunkRegistry = {
     'settings/currencies': () => import('./locales/chunks/en/settings/currencies.json'),
     'settings/accounts-groups': () => import('./locales/chunks/en/settings/accounts-groups.json'),
     'settings/data-management': () => import('./locales/chunks/en/settings/data-management.json'),
-    'settings/preferences': () => import('./locales/chunks/en/settings/preferences.json'),
+    'settings/appearance': () => import('./locales/chunks/en/settings/appearance.json'),
+    'settings/language': () => import('./locales/chunks/en/settings/language.json'),
+    'settings/statistics': () => import('./locales/chunks/en/settings/statistics.json'),
     'settings/ai': () => import('./locales/chunks/en/settings/ai.json'),
     'settings/security': () => import('./locales/chunks/en/settings/security.json'),
     'settings/admin': () => import('./locales/chunks/en/settings/admin.json'),
@@ -102,7 +104,9 @@ const chunkRegistry: ChunkRegistry = {
     'settings/currencies': () => import('./locales/chunks/uk/settings/currencies.json'),
     'settings/accounts-groups': () => import('./locales/chunks/uk/settings/accounts-groups.json'),
     'settings/data-management': () => import('./locales/chunks/uk/settings/data-management.json'),
-    'settings/preferences': () => import('./locales/chunks/uk/settings/preferences.json'),
+    'settings/appearance': () => import('./locales/chunks/uk/settings/appearance.json'),
+    'settings/language': () => import('./locales/chunks/uk/settings/language.json'),
+    'settings/statistics': () => import('./locales/chunks/uk/settings/statistics.json'),
     'settings/ai': () => import('./locales/chunks/uk/settings/ai.json'),
     'settings/security': () => import('./locales/chunks/uk/settings/security.json'),
     'settings/admin': () => import('./locales/chunks/uk/settings/admin.json'),
