+++ packages/frontend/src/i18n/locales/chunks/en/settings/index.json
@@ -6,12 +6,16 @@
       "tags": "Tags",
       "accountsGroups": "Accounts groups",
       "importData": "Import Data",
-      "preferences": "Preferences",
       "ai": "AI",
       "security": "Security",
       "admin": "Admin",
       "sharedWithMe": "Shared with me",
-      "household": "Household"
+      "household": "Household",
+      "personalGroup": "Personal",
+      "workspaceGroup": "Workspace",
+      "appearance": "Appearance",
+      "language": "Language",
+      "statistics": "Statistics"
     },
     "backToSettings": "Back to Settings"
   }
