--- packages/frontend/src/i18n/locales/chunks/en/settings/preferences.json
@@ -1,37 +1,0 @@
-{
-  "settings": {
-    "preferences": {
-      "title": "Preferences",
-      "description": "Customize your app experience",
-      "theme": {
-        "title": "Theme",
-        "description": "Choose how the app looks. System matches your device setting."
-      },
-      "language": {
-        "title": "Language",
-        "description": "Display language for the app interface."
-      },
-      "sidebar": {
-        "title": "Sidebar sections",
-        "description": "Pick which sections appear in the sidebar.",
-        "alwaysVisibleTooltip": "This section is always visible."
-      },
-      "quickStart": {
-        "title": "Quick Start Guide",
-        "description": "The Quick Start guide helps you learn about the app's features. You can reopen it anytime if you previously dismissed it.",
-        "button": "Reopen Quick Start",
-        "activeLabel": "Currently active"
-      },
-      "balanceStats": {
-        "title": "Balance & Statistics",
-        "description": "Control how your financial data is calculated and displayed across the app.",
-        "creditLimit": {
-          "label": "Include credit limits in balance",
-          "description": "When enabled, credit limits are subtracted from account balances in statistics and the balance trend chart. This shows your actual financial position instead of available credit.",
-          "successNotification": "Setting updated successfully",
-          "errorNotification": "Failed to update setting"
-        }
-      }
-    }
-  }
-}
