+++ packages/frontend/src/i18n/locales/chunks/en/settings/statistics.json
@@ -1,0 +1,14 @@
+{
+  "settings": {
+    "statistics": {
+      "title": "Statistics",
+      "description": "Control how your financial data is calculated and displayed across the app.",
+      "creditLimit": {
+        "label": "Include credit limits in balance",
+        "description": "When enabled, credit limits are subtracted from account balances in statistics and the balance trend chart. This shows your actual financial position instead of available credit.",
+        "successNotification": "Setting updated successfully",
+        "errorNotification": "Failed to update setting"
+      }
+    }
+  }
+}
