+++ packages/frontend/src/i18n/locales/chunks/uk/settings/appearance.json
@@ -1,0 +1,17 @@
+{
+  "settings": {
+    "appearance": {
+      "title": "Зовнішній вигляд",
+      "description": "Налаштуйте вигляд застосунку та які розділи відображаються в бічній панелі.",
+      "theme": {
+        "title": "Тема",
+        "description": "Виберіть зовнішній вигляд застосунку. «Системна» відповідає налаштуванню пристрою."
+      },
+      "sidebar": {
+        "title": "Секції бічної панелі",
+        "description": "Виберіть, які секції відображати в бічній панелі.",
+        "alwaysVisibleTooltip": "Ця секція завжди видима."
+      }
+    }
+  }
+}
