+++ packages/frontend/src/i18n/locales/chunks/uk/settings/language.json
@@ -1,0 +1,8 @@
+{
+  "settings": {
+    "language": {
+      "title": "Мова",
+      "description": "Мова інтерфейсу застосунку."
+    }
+  }
+}
