--- packages/frontend/src/i18n/locales/chunks/uk/settings/preferences.json
@@ -1,37 +1,0 @@
-{
-  "settings": {
-    "preferences": {
-      "title": "Налаштування",
-      "description": "Налаштуйте свій досвід користування додатком",
-      "theme": {
-        "title": "Тема",
-        "description": "Виберіть зовнішній вигляд застосунку. «Системна» відповідає налаштуванню пристрою."
-      },
-      "language": {
-        "title": "Мова",
-        "description": "Мова інтерфейсу застосунку."
-      },
-      "sidebar": {
-        "title": "Секції бічної панелі",
-        "description": "Виберіть, які секції відображати в бічній панелі.",
-        "alwaysVisibleTooltip": "Ця секція завжди видима."
-      },
-      "quickStart": {
-        "title": "Посібник Quick Start",
-        "description": "Посібник Quick Start допомагає вам дізнатися про функції додатку. Ви можете відкрити його будь-коли, якщо раніше приховали.",
-        "button": "Відкрити Quick Start",
-        "activeLabel": "Наразі активний"
-      },
-      "balanceStats": {
-        "title": "Баланс і статистика",
-        "description": "Налаштуйте спосіб розрахунку та відображення фінансових даних у додатку.",
-        "creditLimit": {
-          "label": "Враховувати кредитний ліміт у балансі",
-          "description": "Коли увімкнено, кредитні ліміти віднімаються від балансів рахунків у статистиці та графіку балансу. Це показує вашу реальну фінансову позицію замість доступного кредиту.",
-          "successNotification": "Налаштування успішно оновлено",
-          "errorNotification": "Не вдалося оновити налаштування"
-        }
-      }
-    }
-  }
-}
