+++ packages/frontend/src/i18n/locales/chunks/uk/settings/statistics.json
@@ -1,0 +1,14 @@
+{
+  "settings": {
+    "statistics": {
+      "title": "Статистика",
+      "description": "Налаштуйте спосіб розрахунку та відображення фінансових даних у застосунку.",
+      "creditLimit": {
+        "label": "Враховувати кредитні ліміти в балансі",
+        "description": "Коли увімкнено, кредитні ліміти віднімаються від балансів рахунків у статистиці та графіку балансу. Це показує вашу реальну фінансову позицію замість доступного кредиту.",
+        "successNotification": "Налаштування успішно оновлено",
+        "errorNotification": "Не вдалося оновити налаштування"
+      }
+    }
+  }
+}
