+++ packages/frontend/src/i18n/types.ts
@@ -43,7 +43,9 @@ const I18N_CHUNKS = [
   'settings/currencies',
   'settings/accounts-groups',
   'settings/data-management',
-  'settings/preferences',
+  'settings/appearance',
+  'settings/language',
+  'settings/statistics',
   'settings/ai',
   'settings/security',
   'settings/admin',
