+++ packages/frontend/src/routes/constants.ts
@@ -43,7 +43,9 @@ export const ROUTES_NAMES = Object.freeze({
   settingsTags: `${ROUTER_LAYOUTS.dashboard}.settings.tags`,
   settingsAccounts: `${ROUTER_LAYOUTS.dashboard}.settings.accounts-groups`,
   settingsDataManagement: `${ROUTER_LAYOUTS.dashboard}.settings.data-management`,
-  settingsPreferences: `${ROUTER_LAYOUTS.dashboard}.settings.preferences`,
+  settingsAppearance: `${ROUTER_LAYOUTS.dashboard}.settings.appearance`,
+  settingsLanguage: `${ROUTER_LAYOUTS.dashboard}.settings.language`,
+  settingsStatistics: `${ROUTER_LAYOUTS.dashboard}.settings.statistics`,
   settingsAi: `${ROUTER_LAYOUTS.dashboard}.settings.ai`,
   settingsAiFeatures: `${ROUTER_LAYOUTS.dashboard}.settings.ai.features`,
   settingsAiKeys: `${ROUTER_LAYOUTS.dashboard}.settings.ai.keys`,
