+++ packages/frontend/src/routes/index.ts
@@ -236,7 +236,9 @@ const routes: RouteRecordRaw[] = [
             'settings/currencies',
             'settings/accounts-groups',
             'settings/data-management',
-            'settings/preferences',
+            'settings/appearance',
+            'settings/language',
+            'settings/statistics',
             'settings/ai',
             'settings/security',
             'settings/admin',
@@ -275,10 +277,22 @@ const routes: RouteRecordRaw[] = [
             meta: { i18nChunks: ['settings/data-management'] as I18nChunkName[] },
           },
           {
-            path: 'preferences',
-            name: ROUTES_NAMES.settingsPreferences,
-            component: () => import('@/pages/settings/subpages/preferences/index.vue'),
-            meta: { i18nChunks: ['settings/preferences'] as I18nChunkName[] },
+            path: 'appearance',
+            name: ROUTES_NAMES.settingsAppearance,
+            component: () => import('@/pages/settings/subpages/appearance/index.vue'),
+            meta: { i18nChunks: ['settings/appearance'] as I18nChunkName[] },
+          },
+          {
+            path: 'language',
+            name: ROUTES_NAMES.settingsLanguage,
+            component: () => import('@/pages/settings/subpages/language/index.vue'),
+            meta: { i18nChunks: ['settings/language'] as I18nChunkName[] },
+          },
+          {
+            path: 'statistics',
+            name: ROUTES_NAMES.settingsStatistics,
+            component: () => import('@/pages/settings/subpages/statistics/index.vue'),
+            meta: { i18nChunks: ['settings/statistics'] as I18nChunkName[] },
           },
           {
             path: 'admin',
